Export/vp 297 move providers to separate projects (#532)
