+++ VirtoCommerce.ExportModule.CsvProvider/EmptyClassMapImpl.cs
@@ -1,0 +1,6 @@
+using CsvHelper.Configuration;
+
+namespace VirtoCommerce.ExportModule.CsvProvider
+{
+    public class EmptyClassMapImpl<T> : ClassMap<T> { }
+}
