+++ VirtoCommerce.ExportModule.CsvProvider/MetadataFilteredMap.cs
@@ -1,83 +1,15 @@
 using System;
 using System.Collections.Generic;
 using System.Globalization;
-using System.IO;
 using System.Linq;
 using System.Reflection;
-using CsvHelper;
 using CsvHelper.Configuration;
 using VirtoCommerce.ExportModule.Core.Model;
-using VirtoCommerce.ExportModule.Core.Services;
 using VirtoCommerce.ExportModule.Data.Extensions;
-using VirtoCommerce.ExportModule.Data.Model;
 using VirtoCommerce.Platform.Core.Common;
 
-namespace VirtoCommerce.ExportModule.Data.Services
+namespace VirtoCommerce.ExportModule.CsvProvider
 {
-    public sealed class CsvExportProvider : IExportProvider
-    {
-        public string TypeName => nameof(CsvExportProvider);
-        public ExportedTypePropertyInfo[] IncludedProperties { get; private set; }
-        public string ExportedFileExtension => "csv";
-        public bool IsTabular => true;
-        public IExportProviderConfiguration Configuration { get; }
-
-        private CsvWriter _csvWriter;
-
-        public CsvExportProvider(ExportDataRequest exportDataRequest)
-        {
-            if (exportDataRequest == null)
-            {
-                throw new ArgumentNullException(nameof(exportDataRequest));
-            }
-
-            Configuration = exportDataRequest.ProviderConfig;
-            IncludedProperties = exportDataRequest.DataQuery?.IncludedProperties;
-        }
-
-        public void WriteRecord(TextWriter writer, IExportable objectToRecord)
-        {
-            EnsureWriterCreated(writer);
-
-            AddClassMap(objectToRecord.GetType());
-
-            _csvWriter.WriteRecords(new object[] { objectToRecord });
-        }
-
-        public void Dispose()
-        {
-            _csvWriter?.Flush();
-            _csvWriter?.Dispose();
-        }
-
-
-        private void EnsureWriterCreated(TextWriter textWriter)
-        {
-            if (_csvWriter == null)
-            {
-                var csvConfiguration = (Configuration as CsvProviderConfiguration)?.Configuration ?? new Configuration();
-
-                _csvWriter = new CsvWriter(textWriter, csvConfiguration, true);
-            }
-        }
-
-        private void AddClassMap(Type objectType)
-        {
-            var csvConfiguration = _csvWriter.Configuration;
-            var mapForType = csvConfiguration.Maps[objectType];
-
-            if (mapForType == null)
-            {
-                var constructor = typeof(MetadataFilteredMap<>).MakeGenericType(objectType).GetConstructor(IncludedProperties != null
-                    ? new[] { typeof(ExportedTypePropertyInfo[]) }
-                    : Array.Empty<Type>());
-                var classMap = (ClassMap)constructor.Invoke(IncludedProperties != null ? new[] { IncludedProperties } : null);
-
-                csvConfiguration.RegisterClassMap(classMap);
-            }
-        }
-    }
-
     /// <summary>
     /// Custom ClassMap implementation which includes type properties. Supports nested properties.
     /// Does not map <see cref="IEnumerable<Entity>"/> as these are not representable in CSV structure in suitable manner.
@@ -171,6 +103,4 @@ namespace VirtoCommerce.ExportModule.Data.Services
             return memberMap;
         }
     }
-
-    public class EmptyClassMapImpl<T> : ClassMap<T> { }
 }
