+++ VirtoCommerce.ExportModule.CsvProvider/Properties/AssemblyInfo.cs
@@ -1,0 +1,5 @@
+using System.Reflection;
+
+[assembly: AssemblyTitle("VirtoCommerce.ExportModule.CsvProvider")]
+[assembly: AssemblyDescription("")]
+
