+++ VirtoCommerce.ExportModule.Data/packages.config
@@ -3,7 +3,6 @@
   <package id="CacheManager.Core" version="0.8.0" targetFramework="net461" />
   <package id="Common.Logging" version="3.3.1" targetFramework="net461" />
   <package id="Common.Logging.Core" version="3.3.1" targetFramework="net461" />
-  <package id="CsvHelper" version="12.1.2" targetFramework="net461" />
   <package id="EntityFramework" version="6.2.0" targetFramework="net461" />
   <package id="Newtonsoft.Json" version="11.0.2" targetFramework="net461" />
   <package id="System.ValueTuple" version="4.4.0" targetFramework="net461" />
