+++ VirtoCommerce.ExportModule.JsonProvider/JsonExportProvider.cs
@@ -1,14 +1,11 @@
 using System;
 using System.IO;
-using System.Linq;
 using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 using VirtoCommerce.ExportModule.Core.Model;
 using VirtoCommerce.ExportModule.Core.Services;
-using VirtoCommerce.ExportModule.Data.Model;
 using VirtoCommerce.Platform.Core.Common;
 
-namespace VirtoCommerce.ExportModule.Data.Services
+namespace VirtoCommerce.ExportModule.JsonProvider
 {
     public sealed class JsonExportProvider : IExportProvider
     {
@@ -16,6 +13,8 @@ namespace VirtoCommerce.ExportModule.Data.Services
         public ExportedTypePropertyInfo[] IncludedProperties { get; private set; }
         public string ExportedFileExtension => "json";
         public bool IsTabular => false;
+        [JsonIgnore]
+        // TODO: Temporary, before config rework - need to store only specific properties, not whole huge provider specific configs
         public IExportProviderConfiguration Configuration { get; }
 
         private readonly JsonSerializer _serializer;
@@ -67,49 +66,6 @@ namespace VirtoCommerce.ExportModule.Data.Services
                 _jsonTextWriter = new JsonTextWriter(writer);
                 _jsonTextWriter.CloseOutput = false;
                 _jsonTextWriter.WriteStartArray();
-            }
-        }
-    }
-
-    public class ObjectDiscriminatorJsonConverter : JsonConverter
-    {
-        private readonly Type[] _types;
-        private readonly JsonSerializer _jsonSerializer;
-
-        public override bool CanRead => false;
-        public override bool CanWrite => true;
-
-        public ObjectDiscriminatorJsonConverter(JsonSerializerSettings jsonSerializerSettings, params Type[] types)
-        {
-            _types = types;
-            _jsonSerializer = JsonSerializer.Create(jsonSerializerSettings);
-        }
-
-        public override bool CanConvert(Type objectType)
-        {
-            return _types.Any(x => x.IsAssignableFrom(objectType));
-        }
-
-        public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
-        {
-            // Here we use another serializer with the same settings to avoid infinity cycle
-            var jToken = JToken.FromObject(value, _jsonSerializer);
-
-            if (jToken.Type == JTokenType.Object)
-            {
-                var jObject = (JObject)jToken;
-
-                jObject.AddFirst(new JProperty("$discriminator", value.GetType().FullName));
-                jObject.WriteTo(writer);
-            }
-            else
-            {
-                jToken.WriteTo(writer);
             }
         }
     }
