+++ VirtoCommerce.ExportModule.JsonProvider/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Newtonsoft.Json" version="11.0.2" targetFramework="net461" />
+  <package id="VirtoCommerce.Platform.Core" version="2.13.44" targetFramework="net461" />
+</packages>
