+++ VirtoCommerce.ExportModule.Tests/JsonExportProviderTests.cs
@@ -7,8 +7,7 @@ using System.Threading.Tasks;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using VirtoCommerce.ExportModule.Core.Model;
-using VirtoCommerce.ExportModule.Data.Model;
-using VirtoCommerce.ExportModule.Data.Services;
+using VirtoCommerce.ExportModule.JsonProvider;
 using Xunit;
 
 namespace VirtoCommerce.ExportModule.Tests
